@@ -16,40 +16,41 @@
     "start": "node dist/server.js"
   },
   "devDependencies": {
-    "@babel/core": "^7.15.8",
-    "@storybook/addon-actions": "^6.4.0-beta.19",
-    "@storybook/addon-essentials": "^6.4.0-beta.19",
-    "@storybook/addon-links": "^6.4.0-beta.19",
-    "@storybook/react": "^6.4.0-beta.19",
+    "@babel/core": "^7.17.10",
+    "@storybook/addon-actions": "^6.5.0-beta.1",
+    "@storybook/addon-essentials": "^6.5.0-beta.1",
+    "@storybook/addon-links": "^6.5.0-beta.1",
+    "@storybook/builder-vite": "^0.1.33",
+    "@storybook/react": "^6.5.0-beta.1",
     "@types/express": "^4.17.13",
-    "@types/node": "^16.11.4",
-    "@types/react": "^17.0.32",
-    "@types/react-dom": "^17.0.10",
-    "@types/react-router-dom": "^5.3.1",
-    "@typescript-eslint/eslint-plugin": "^5.1.0",
-    "@typescript-eslint/parser": "^5.1.0",
+    "@types/node": "^17.0.31",
+    "@types/react": "^18.0.8",
+    "@types/react-dom": "^18.0.3",
+    "@types/react-router-dom": "^5.3.3",
+    "@typescript-eslint/eslint-plugin": "^5.22.0",
+    "@typescript-eslint/parser": "^5.22.0",
     "@vitejs/plugin-react-refresh": "^1.3.6",
-    "babel-loader": "^8.2.3",
-    "concurrently": "^6.3.0",
-    "eslint": "^8.1.0",
-    "eslint-config-prettier": "^8.3.0",
-    "http-proxy-middleware": "^2.0.1",
+    "babel-loader": "^8.2.5",
+    "concurrently": "^7.1.0",
+    "eslint": "^8.14.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-storybook": "^0.5.11",
+    "http-proxy-middleware": "^2.0.6",
     "husky": "^7.0.4",
-    "lint-staged": "^11.2.4",
-    "prettier": "2.4.1",
-    "storybook-builder-vite": "^0.1.2",
-    "stylelint": "^14.0.0",
+    "lint-staged": "^12.4.1",
+    "prettier": "2.6.2",
+    "stylelint": "^14.8.1",
     "stylelint-config-prettier": "^9.0.3",
-    "stylelint-config-standard": "^23.0.0",
+    "stylelint-config-standard": "^25.0.0",
     "ts-node-dev": "^1.1.8",
-    "typescript": "^4.4.4",
-    "vite": "^2.6.10"
+    "typescript": "^4.6.4",
+    "vite": "^2.9.7"
   },
   "dependencies": {
-    "dotenv": "^10.0.0",
-    "express": "^4.17.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0"
+    "dotenv": "^16.0.0",
+    "express": "^4.18.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-router-dom": "^6.3.0"
   }
 }
